Refactor role guard and user models
This commit refactors the role guard in `rol.guard.tsx` to use the `useDecoded` hook for retrieving the user's role. It also updates the `Props` interface to use `role` instead of `rol`. Additionally, the commit removes the `PRIVATE` route from `PrivateRoutes` in `routes.ts` and adds new user models `UserLogin` and `UserRegister` in `user.model.ts`. Finally, the commit updates the `loginService` and `registerService` functions in `auth.service.ts` to use the new user models.